relocate use-toastr-on-error.tsx
