+++ apps/app/src/client/services/create-page/index.ts
@@ -1,3 +1,2 @@
 export * from './use-create-page-and-transit';
 export * from './use-create-template-page';
-export * from './use-toastr-on-error';
