+++ apps/app/src/client/services/use-toastr-on-error.tsx
